Add check for node.body in referencer
Add comment explaining the node.body check

Allow tests to run with babel

We will remove this soon but need to add tests for new features
in the mean time.

Add typescript-eslint-parser as dev dependency

This used in tests for typescript features.

Strict scope should be false when there is no body

This fixes an issue with typescript having an empty body
for some functions

Add tests for typescript scope analysis

This includes a test for multiple call signatures.

Add comment explaining the node.body check

Allow tests to run with babel

We will remove this soon but need to add tests for new features
in the mean time.

Strict scope should be false when there is no body

This fixes an issue with typescript having an empty body
for some functions

Add tests for typescript scope analysis

This includes a test for multiple call signatures.